+++ demo/starter/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <!-- head -->
+</head>
+<body>
+  <div id="app"></div>
+  <script type="module" src="__ENTRY__"></script>
+  <div id="mermaid-rendering-container"></div>
+  <!-- body -->
+</body>
+</html>
